requirements.txt
Django==4.2.6
django-multiselectfield==0.1.12
Pillow==10.1.0
channels==4.0.0
websocket==0.2.1
daphne==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
